activate user from app api call with email token, issue access token

+++ src/Controller/ApiController.php
@@ -2,14 +2,19 @@
 
 namespace Bone\BoneUserApi\Controller;
 
+use Bone\BoneDoctrine\EntityManagerAwareInterface;
+use Bone\BoneDoctrine\HasEntityManagerTrait;
 use Bone\Controller\Controller;
-use Bone\Http\Response\HtmlResponse;
 use Bone\I18n\Form;
 use Bone\Mail\EmailMessage;
 use Bone\Mail\Service\MailService;
+use Bone\OAuth2\Controller\AuthServerController;
+use Bone\OAuth2\Entity\Client;
+use Bone\Server\SessionAwareInterface;
+use Bone\Server\Traits\HasSessionTrait;
 use Bone\User\Form\PersonForm;
-use Bone\User\Form\RegistrationForm;
 use DateTime;
+use Defuse\Crypto\Key;
 use Del\Entity\User;
 use Del\Exception\EmailLinkException;
 use Del\Exception\UserException;
@@ -17,26 +22,39 @@ use Del\Factory\CountryFactory;
 use Del\Form\Field\Text\EmailAddress;
 use Del\Service\UserService;
 use Del\Value\User\State;
+use Exception;
+use GuzzleHttp\Psr7\ServerRequest;
 use Laminas\Diactoros\Response\JsonResponse;
-use Laminas\Diactoros\Response\RedirectResponse;
+use Laminas\Diactoros\ServerRequestFactory;
+use Laminas\Diactoros\Uri;
+use League\OAuth2\Server\CryptKey;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
-class ApiController extends Controller
+class ApiController extends Controller implements EntityManagerAwareInterface, SessionAwareInterface
 {
-    private UserService $userService;
-    private MailService $mailService;
+    use HasEntityManagerTrait;
+    use HasSessionTrait;
+
     private array $nativeAppSettings;
+    private array $oauth2settings;
+    private AuthServerController $authServerController;
+    private CryptKey $privateKey;
+    private Key $encryptionKey;
+    private MailService $mailService;
+    private UserService $userService;
+    private string $verifier = 'RaYFJonXO4pxNh7w5ygj90D8ruMPvwigU.BPMNp.O9~-oym396zbL0Lf9wGPPJSWkg1Lrz.4rHgbn_XwoyapfcTfXIONjhlOcA78y11PQIt1ueFlCbJ6DZ892G3HpLkt';
 
     /**
      * BoneUserController constructor.
      * @param UserService $userService
      */
-    public function __construct(UserService $userService, MailService $mailService, array $nativeAppSettings)
+    public function __construct(UserService $userService, MailService $mailService, AuthServerController $authServerController, array $nativeAppSettings)
     {
         $this->userService = $userService;
         $this->mailService = $mailService;
         $this->nativeAppSettings = $nativeAppSettings;
+        $this->authServerController = $authServerController;
     }
 
     /**
@@ -97,8 +115,8 @@ class ApiController extends Controller
      * )
      * @param ServerRequestInterface $request
      * @param array $args
-     * @throws \Exception
      * @return ResponseInterface
+     * @throws Exception
      */
     public function registerAction(ServerRequestInterface $request, array $args): ResponseInterface
     {
@@ -163,11 +181,21 @@ class ApiController extends Controller
      *                     type="string",
      *                     example="fake@email.com",
      *                     description="The account to activate"
-     *                 ),@OA\Property(
+     *                 ), @OA\Property(
      *                     property="token",
      *                     type="string",
      *                     example="xxxxxxxxxx",
-     *                     description="The security token"
+     *                     description="The security token from the email"
+     *                 ), @OA\Property(
+     *                     property="clientId",
+     *                     type="string",
+     *                     example="xxxxxxxxxx",
+     *                     description="The clientId of the app"
+     *                 ), @OA\Property(
+     *                     property="password",
+     *                     type="string",
+     *                     example="xxxxxxxxxx",
+     *                     description="The new users password"
      *                 )
      *             )
      *         )
@@ -185,26 +213,90 @@ class ApiController extends Controller
      */
     public function activateAction(ServerRequestInterface $request): ResponseInterface
     {
-        $email = $request->getAttribute('email');
-        $token = $request->getAttribute('token');
+        $body = $request->getParsedBody();
+        $clientId = $body['clientId'];
+        $email = $body['email'];
+        $token = $body['token'];
+        $password = $body['password'];
         $userService = $this->userService;
+        $client = $this->getEntityManager()->getRepository(Client::class)->getClientEntity($clientId);
+
+        if (!$client) {
+            return new JsonResponse([
+                'error' => 'Client not found'
+            ], 404);
+        }
 
         try {
             $link = $userService->findEmailLink($email, $token);
             $user = $link->getUser();
             $user->setState(new State(State::STATE_ACTIVATED));
             $user->setLastLogin(new DateTime());
+            $userService->changePassword($user, $password);
             $userService->saveUser($user);
             $userService->deleteEmailLink($link);
-            /** @todo create an access token */
-            $body = ['access_token' => 'xxx'];
-            $status = 200;
+            $this->getSession()->set('user', $user->getId());
+            $authRequest = $this->getAuthRequest($clientId);
+            $response = $this->authServerController->authorizeAction($authRequest, []);
+            $redirectUri = $response->getHeaderLine('Location');
+            $url = \parse_url($redirectUri);\parse_str($url['query'], $params);
+            $code = $params['code'];
+            $tokenRequest = $this->getTokenRequest($clientId, $code);
+            $response = $this->authServerController->accessTokenAction($tokenRequest, []);
+
+            return $response;
+
         } catch (EmailLinkException $e) {
             $body = ['error' => $e->getMessage()];
             $status = 400;
+        } catch (Exception $e) {
+            return new JsonResponse(['error' => $e->getMessage()]);
         }
 
         return new JsonResponse($body, $status);
+    }
+
+    private function getAuthRequest(string $clientId): ServerRequestInterface
+    {
+        $url = $this->getSiteConfig()->getBaseUrl() . '/oauth2/authorize';
+        $query = [
+            'client_id' => $clientId,
+            'code_challenge' => $this->generatePkceCodeChallenge($this->verifier),
+            'code_challenge_method' => 'S256',
+            'response_type' => 'code',
+            'redirect_uri' => $this->nativeAppSettings['deepLink'] . 'oauth2/callback',
+            'scope' => 'basic',
+            'state' => 'xxx',
+        ];
+
+        $request = new ServerRequest('POST', $url);
+        $request = $request->withQueryParams($query);
+
+        return $request;
+    }
+
+    private function getTokenRequest(string $clientId, string $code): ServerRequestInterface
+    {
+        $url = $this->getSiteConfig()->getBaseUrl() . '/oauth2/token';
+        $body = [
+            'grant_type' => 'authorization_code',
+            'client_id' => $clientId,
+            'code_verifier' => $this->verifier,
+            'redirect_uri' => $this->nativeAppSettings['deepLink'] . 'oauth2/callback',
+            'code' => $code
+        ];
+
+        $request = ServerRequestFactory::fromGlobals(null, [], $body);
+        $request = $request->withMethod('POST');
+        $request = $request->withHeader('Content-Type', 'application/x-www-form-urlencoded');
+
+        return $request->withUri(new Uri($url));
+    }
+
+    private function generatePkceCodeChallenge(string $verifier)
+    {
+        $hash = hash('sha256', $verifier, true);
+        return rtrim(strtr(base64_encode($hash), '+/', '-_'), '=');
     }
 
     /**
@@ -267,8 +359,8 @@ class ApiController extends Controller
      * )
      * @param ServerRequestInterface $request
      * @param array $args
-     * @throws \Exception
      * @return ResponseInterface
+     * @throws Exception
      */
     public function editProfileAction(ServerRequestInterface $request, array $args): ResponseInterface
     {
